feat: added more tests, working on responses
