+++ packages/ts-to-openapi/src/responses.ts
@@ -1,19 +1,99 @@
 import type { OpenAPIV3_1 } from 'openapi-types'
 import {
+  type Expression,
   type Node,
   type Program,
+  type ReturnStatement,
+  SyntaxKind,
+  isArrowFunction,
+  isBlock,
   isCallExpression,
+  isCaseClause,
+  isConstructorDeclaration,
+  isDefaultClause,
+  isForInStatement,
+  isForOfStatement,
+  isForStatement,
+  isFunctionDeclaration,
+  isFunctionExpression,
   isIdentifier,
+  isIfStatement,
+  isMethodDeclaration,
   isObjectLiteralExpression,
   isPropertyAccessExpression,
   isReturnStatement,
+  isSwitchStatement,
+  isTryStatement,
+  isWhileStatement,
 } from 'typescript'
 
-import { getReturnStatements } from './return-statements'
 import { getSchemaFromTypeNode } from './type-nodes'
 import type { FileNameResolver } from './types'
 
-/** Convert the generator results to response objects */
+/**
+ * Generator to grab all nested return statements from a node
+ *
+ * @see https://stackoverflow.com/a/76551960
+ */
+export function* getReturnStatements(
+  node: Node,
+): Generator<ReturnStatement | Expression> {
+  // Got the return statement
+  if (isReturnStatement(node)) {
+    yield node
+  } else if (isBlock(node) || isCaseClause(node) || isDefaultClause(node)) {
+    for (const stmt of node.statements) {
+      yield* getReturnStatements(stmt)
+    }
+  } else if (isIfStatement(node)) {
+    yield* getReturnStatements(node.thenStatement)
+    if (node.elseStatement) {
+      yield* getReturnStatements(node.elseStatement)
+    }
+  } else if (
+    isForStatement(node) ||
+    isForOfStatement(node) ||
+    isForInStatement(node) ||
+    isWhileStatement(node)
+  ) {
+    yield* getReturnStatements(node.statement)
+  } else if (isSwitchStatement(node)) {
+    for (const clause of node.caseBlock.clauses) {
+      yield* getReturnStatements(clause)
+    }
+  } else if (isTryStatement(node)) {
+    yield* getReturnStatements(node.tryBlock)
+    if (node.catchClause) {
+      yield* getReturnStatements(node.catchClause.block)
+    }
+    if (node.finallyBlock) {
+      yield* getReturnStatements(node.finallyBlock)
+    }
+  } else if (
+    isMethodDeclaration(node) ||
+    isFunctionDeclaration(node) ||
+    isArrowFunction(node) ||
+    isFunctionExpression(node) ||
+    isConstructorDeclaration(node)
+  ) {
+    if (node.body) {
+      if (isBlock(node.body)) {
+        yield* getReturnStatements(node.body)
+      } else {
+        yield node.body
+      }
+    }
+  }
+}
+
+/**
+ * Convert the generator results to response schema objects
+ *
+ * TODO:
+ * - other types besides json
+ * - grab jsDoc
+ * - pass in a predicate as this if statement is meant for next
+ */
 export const generateResponses = (
   node: Node,
   program: Program,
@@ -23,7 +103,10 @@ export const generateResponses = (
   const statements = Array.from(generator)
 
   return statements.reduce<OpenAPIV3_1.ResponsesObject>((prev, statement) => {
+    console.log(statement.getText())
+
     // Check for a Response.json
+    // we will probably pass in this
     if (
       isReturnStatement(statement) &&
       statement.expression &&
@@ -43,10 +126,12 @@ export const generateResponses = (
 
       // Check the type of the payload
       if (payload) {
+        console.log(JSON.stringify(payload, null, 2))
+
         schema = getSchemaFromTypeNode(payload, program, fileNameResolver)
       }
       // Update the status
-      // TODO headers, statusText
+      // TODO: headers, statusText
       if (resp && isObjectLiteralExpression(resp)) {
         const respSchema = getSchemaFromTypeNode(
           resp.properties[0],
@@ -59,8 +144,7 @@ export const generateResponses = (
       return {
         ...prev,
         [String(status)]: {
-          description:
-            'TODO grab this from jsDoc or use set defaults per status',
+          description: 'TODO: grab this from jsdoc and add a default',
           content: {
             'application/json': {
               schema,
@@ -71,14 +155,15 @@ export const generateResponses = (
           },
         },
       }
-    }
-    // Other types of returns
-    // arrow
-    // new Response with types
-    else {
+    } else {
       console.log('TODO other types of returns')
-
-      return prev
+      return {
+        ...prev,
+        '999': {
+          description: 'unknown return type: ' + SyntaxKind[statement.kind],
+          content: { 'application/json': {} },
+        },
+      }
     }
   }, {})
 }
