+++ packages/ts-to-openapi/src/test-setup.ts
@@ -1,0 +1,53 @@
+import fg from 'fast-glob'
+import { readFileSync } from 'fs'
+import path from 'path'
+import {
+  type CompilerHost,
+  JSDocParsingMode,
+  ScriptKind,
+  ScriptTarget,
+  createProgram,
+  createSourceFile,
+} from 'typescript'
+
+const compilerHost: CompilerHost = {
+  fileExists: () => true,
+  getCanonicalFileName: (filename) => filename,
+  getCurrentDirectory: () => '',
+  getDefaultLibFileName: () => '',
+  getNewLine: () => '\n',
+  getSourceFile: (filename) =>
+    createSourceFile(
+      filename,
+      readFileSync(filename).toString(),
+      ScriptTarget.Latest,
+      false,
+      ScriptKind.TS,
+    ),
+  jsDocParsingMode: JSDocParsingMode.ParseAll,
+  readFile: () => undefined,
+  useCaseSensitiveFileNames: () => true,
+  writeFile: () => null,
+}
+
+const programFileNames = await fg('src/fixtures/*.ts')
+
+// Intialize typescript program
+export const program = createProgram(
+  programFileNames,
+  {
+    noResolve: true,
+    target: ScriptTarget.Latest,
+  },
+  compilerHost,
+)
+
+export const fileResolver = (source: string, target: string) => {
+  const sourceExt = path.extname(source)
+  const targetExt = path.extname(target)
+
+  const targetRelative = target + (targetExt ? '' : sourceExt)
+  const targetPath = path.join(source.replace(/\/([^/]+)$/, ''), targetRelative)
+
+  return targetPath
+}
