+++ packages/ts-to-openapi/src/type-nodes.ts
@@ -202,14 +202,12 @@ export const getSchemaFromTypeNode = (
       }
   }
   // Intersection
-  else if (isIntersectionTypeNode(typeNode)) {
-    console.log(typeNode)
+  else if (isIntersectionTypeNode(typeNode))
     return {
       allOf: typeNode.types.map((type) =>
         getSchemaFromTypeNode(type, program, fileNameResolver),
       ),
     }
-  }
   // Type reference
   else if (isTypeReferenceNode(typeNode) && isIdentifier(typeNode.typeName)) {
     const typeChecker = program.getTypeChecker()
